+++ src/tool/offline_server/server.rs
@@ -98,6 +98,11 @@ pub async fn offline_rpc_state(
     let sync_network_context =
         SyncNetworkContext::new(network_send, peer_manager, state_manager.db_owned());
     let nonce_tracker = NonceTracker::new();
+    let eth_event_handler = Arc::new(EthEventHandler::from_config(
+        &events_config,
+        state_manager.chain_config().eth_chain_id,
+        message_pool.subscriber(),
+    ));
     Ok((
         RPCState {
             state_manager,
@@ -105,7 +110,7 @@ pub async fn offline_rpc_state(
             mpool: message_pool,
             bad_blocks: Default::default(),
             sync_status: Arc::new(ArcSwap::from_pointee(SyncStatusReport::init())),
-            eth_event_handler: Arc::new(EthEventHandler::from_config(&events_config)),
+            eth_event_handler,
             eth_logs_feed: Default::default(),
             sync_network_context,
             start_time: chrono::Utc::now(),
@@ -126,7 +131,7 @@ pub async fn start_offline_server(
     chain: Option<NetworkChain>,
     rpc_port: u16,
     auto_download_snapshot: bool,
-    height: i64,
+    height: ChainEpoch,
     index_backfill_epochs: usize,
     genesis: Option<PathBuf>,
     save_jwt_token: Option<PathBuf>,
@@ -153,8 +158,8 @@ pub async fn start_offline_server(
 
     let inferred_chain = {
         let head = db.heaviest_tipset()?;
-        let genesis = head.genesis(&db)?;
-        NetworkChain::from_genesis_or_devnet_placeholder(genesis.cid())
+        let genesis = head.genesis(db.shallow_clone()).await?;
+        NetworkChain::from_genesis_or_devnet_placeholder(genesis.min_ticket_block().cid())
     };
     let chain = if let Some(chain) = chain {
         anyhow::ensure!(